modify ui structure
